third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-profiling-expected.txt

Tests that CPU profiling works. Bug 52634.

Profiler was enabled.

Running: testProfiling
found pageFunction
found (anonymous function)

Profiler was disabled.
